# Conflicts:
#	app/Http/Controllers/PermissionController.php

+++ app/Http/Controllers/PermissionController.php
@@ -4,11 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Models\Permission;
 use App\Models\Role;
+use App\Permissions;
 use App\Services\Router\Attributes\Get;
 use App\Services\Router\Attributes\Post;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 use Illuminate\Validation\ValidationException;
 use LogicException;
@@ -18,6 +20,9 @@ class PermissionController extends AController
     #[Get('/people-management/permissions', 'permissions')]
     public function default(): View
     {
+        Gate::authorize(Permissions::VIEW_PERMISSION);
+        Gate::authorize(Permissions::VIEW_ROLE);
+
         return view('people-management.permissions', [
             'roles' => Role::with('permissions')->orderBy('role_name')->get(),
             'permissions' => Permission::orderBy('perm_name')->get(),
@@ -30,6 +35,8 @@ class PermissionController extends AController
     #[Post('/people-management/permission', 'create_permission')]
     public function createPermission(Request $request): RedirectResponse
     {
+        Gate::authorize(Permissions::CREATE_PERMISSION);
+
         $data = $request->validate([
             'perm_name' => ['required', 'string', 'max:255', 'unique:permissions,perm_name'],
             'description' => ['nullable', 'string', 'max:255'],
@@ -48,6 +55,8 @@ class PermissionController extends AController
     #[Post('/people-management/role-permission', 'assign_permission')]
     public function assignPermission(Request $request): RedirectResponse
     {
+        Gate::authorize(Permissions::ASSIGN_PERMISSION);
+
         $data = $request->validate([
             'permission_id' => ['required', 'integer', 'exists:permissions,id'],
             'role_ids' => ['sometimes', 'array'],
@@ -69,6 +78,8 @@ class PermissionController extends AController
     #[Post('/people-management/update-permission/{permission}', 'update_permission')]
     public function updatePermission(Request $request, Permission $permission): RedirectResponse
     {
+        Gate::authorize(Permissions::EDIT_PERMISSION);
+
         $data = $request->validate([
             'perm_name' => ['required', 'string', 'max:255',
                 Rule::unique('permissions', 'perm_name')->ignore($permission->id),
@@ -93,6 +104,8 @@ class PermissionController extends AController
     #[Post('/people-management/delete-permission/{permission}', 'delete_permission')]
     public function deletePermission(Permission $permission): RedirectResponse
     {
+        Gate::authorize(Permissions::DELETE_PERMISSION);
+
         $permission->roles()->detach();
 
         $permission->delete();
